chore(workspace-manager): refactor worskspace store to use sveltemap for statuses
