+++ packages/renderer/src/lib/agent-workspaces/AgentWorkspaceCard.svelte
@@ -5,13 +5,13 @@ import { router } from 'tinro';
 
 import { withConfirmation } from '/@/lib/dialogs/messagebox-utils';
 import LoadingIcon from '/@/lib/ui/LoadingIcon.svelte';
-import type { AgentWorkspaceStatus } from '/@/stores/agent-workspaces';
+import type { AgentWorkspaceStatus } from '/@/stores/agent-workspaces.svelte';
 import {
   agentWorkspaceStatuses,
   fetchAgentWorkspaces,
   startAgentWorkspace,
   stopAgentWorkspace,
-} from '/@/stores/agent-workspaces';
+} from '/@/stores/agent-workspaces.svelte';
 import type { AgentWorkspaceSummary } from '/@api/agent-workspace-info';
 
 interface Props {
@@ -20,7 +20,7 @@ interface Props {
 
 let { workspace }: Props = $props();
 
-const status: AgentWorkspaceStatus = $derived($agentWorkspaceStatuses.get(workspace.id) ?? 'stopped');
+const status: AgentWorkspaceStatus = $derived(agentWorkspaceStatuses.get(workspace.id) ?? 'stopped');
 const isRunning = $derived(status === 'running' || status === 'stopping');
 const inProgress = $derived(status === 'starting' || status === 'stopping');
