+++ packages/renderer/src/stores/agent-workspaces.svelte.ts
@@ -16,15 +16,15 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
-import type { Writable } from 'svelte/store';
-import { get, writable } from 'svelte/store';
+import { SvelteMap } from 'svelte/reactivity';
+import { type Writable, writable } from 'svelte/store';
 
 import type { AgentWorkspaceSummary } from '/@api/agent-workspace-info';
 
 export type AgentWorkspaceStatus = 'stopped' | 'running' | 'starting' | 'stopping';
 
 export const agentWorkspaces: Writable<AgentWorkspaceSummary[]> = writable([]);
-export const agentWorkspaceStatuses: Writable<Map<string, AgentWorkspaceStatus>> = writable(new Map());
+export const agentWorkspaceStatuses = new SvelteMap<string, AgentWorkspaceStatus>();
 
 export async function fetchAgentWorkspaces(): Promise<void> {
   const data = await window.listAgentWorkspaces();
@@ -32,40 +32,25 @@ export async function fetchAgentWorkspaces(): Promise<void> {
 }
 
 export async function startAgentWorkspace(id: string): Promise<void> {
-  agentWorkspaceStatuses.update(statuses => {
-    const next = new Map(statuses);
-    next.set(id, 'starting');
-    return next;
-  });
+  agentWorkspaceStatuses.set(id, 'starting');
   try {
     await window.startAgentWorkspace(id);
-    agentWorkspaceStatuses.update(statuses => {
-      const next = new Map(statuses);
-      next.set(id, 'running');
-      return next;
-    });
+    agentWorkspaceStatuses.set(id, 'running');
   } catch (error: unknown) {
-    agentWorkspaceStatuses.update(statuses => {
-      const next = new Map(statuses);
-      next.set(id, 'stopped');
-      return next;
-    });
+    agentWorkspaceStatuses.set(id, 'stopped');
     console.error('Failed to start agent workspace', error);
   }
 }
 
 export async function stopAgentWorkspace(id: string): Promise<void> {
-  const statuses = get(agentWorkspaceStatuses);
-  statuses.set(id, 'stopping');
-  agentWorkspaceStatuses.set(new Map(statuses));
+  agentWorkspaceStatuses.set(id, 'stopping');
   try {
     await window.stopAgentWorkspace(id);
-    statuses.set(id, 'stopped');
+    agentWorkspaceStatuses.set(id, 'stopped');
   } catch (error: unknown) {
-    statuses.set(id, 'running');
+    agentWorkspaceStatuses.set(id, 'running');
     console.error('Failed to stop agent workspace', error);
   }
-  agentWorkspaceStatuses.set(new Map(statuses));
 }
 
 window.addEventListener('system-ready', () => {
